Add constants configuration and refactor services to use centralized constants
This commit introduces a centralized constants module to manage application-wide configuration values. Key changes include:
- Moving hardcoded timeouts, rate limits, and other configuration values to src/constants.ts
- Refactoring BlockchainService, DepositMonitor, TaskProcessor and other services to use constants
- Updating IPFS gateway configuration and reward system parameters
- Improving type safety with const assertions for constant objects

The refactoring improves maintainability by centralizing configuration values and reducing hardcoded magic numbers throughout the codebase.
