+++ bots/kasumi-3/src/services/BlockchainService.ts
@@ -2,6 +2,7 @@ import { Contract, Wallet, ethers } from 'ethers';
 import { IBlockchainService } from '../types';
 import { log } from '../log';
 import { expretry, generateCommitment } from '../utils';
+import { BLOCKCHAIN } from '../constants';
 import ArbiusAbi from '../abis/arbius.json';
 import ArbiusRouterAbi from '../abis/arbiusRouter.json';
 import ERC20Abi from '../abis/erc20.json';
@@ -14,17 +15,11 @@ export class BlockchainService implements IBlockchainService {
   private token: Contract;
   private rpcUrls: string[];
   private nonceCache: { nonce: number; timestamp: number } | null = null;
-  private readonly NONCE_CACHE_TTL = 5000; // 5 seconds
-  private readonly GAS_BUFFER_PERCENT = parseInt(process.env.GAS_BUFFER_PERCENT || '20'); // 20% buffer default
+  private readonly NONCE_CACHE_TTL = BLOCKCHAIN.NONCE_CACHE_TTL;
+  private readonly GAS_BUFFER_PERCENT = parseInt(process.env.GAS_BUFFER_PERCENT || String(BLOCKCHAIN.GAS_BUFFER_PERCENT_DEFAULT));
 
   // Fallback gas limits (used when estimation fails)
-  private readonly FALLBACK_GAS_LIMITS = {
-    submitTask: 200_000n,
-    signalCommitment: 450_000n,
-    submitSolution: 500_000n,
-    approve: 100_000n,
-    validatorDeposit: 150_000n,
-  };
+  private readonly FALLBACK_GAS_LIMITS = BLOCKCHAIN.FALLBACK_GAS_LIMITS;
 
   constructor(
     rpcUrl: string,
@@ -64,6 +59,10 @@ export class BlockchainService implements IBlockchainService {
     this.token = new Contract(tokenAddress, ERC20Abi, this.wallet);
   }
 
+  /**
+   * Get the wallet address used by this blockchain service
+   * @returns The Ethereum address of the wallet
+   */
   getWalletAddress(): string {
     return this.wallet.address;
   }
@@ -134,7 +133,7 @@ export class BlockchainService implements IBlockchainService {
    */
   private async executeTransaction(
     txFunction: (nonce: number) => Promise<ethers.ContractTransactionResponse>,
-    maxRetries: number = 3
+    maxRetries: number = BLOCKCHAIN.NONCE_RETRY_MAX
   ): Promise<ethers.ContractTransactionResponse> {
     let lastError: Error | null = null;
 
@@ -175,6 +174,10 @@ export class BlockchainService implements IBlockchainService {
     throw new Error(`Transaction failed after ${maxRetries} attempts: ${lastError?.message}`);
   }
 
+  /**
+   * Get the AIUS token balance of the wallet
+   * @returns Balance in wei (smallest unit of AIUS)
+   */
   async getBalance(): Promise<bigint> {
     return await expretry('getBalance', async () =>
       await this.token.balanceOf(this.wallet.address)
@@ -225,14 +228,14 @@ export class BlockchainService implements IBlockchainService {
       const balance = await this.getBalance();
       const shortfall = validatorMinimum - validatorStaked;
 
-      // Add 10% buffer to account for potential validator minimum increases
-      const bufferMultiplier = 1.1;
+      // Add buffer to account for potential validator minimum increases
+      const bufferMultiplier = 1 + (BLOCKCHAIN.STAKE_BUFFER_PERCENT / 100);
       const requiredBalance = shortfall * BigInt(Math.floor(bufferMultiplier * 100)) / 100n;
 
       if (balance < requiredBalance) {
         throw new Error(
           `Insufficient balance to stake. Need ${ethers.formatEther(requiredBalance)} AIUS ` +
-          `(shortfall: ${ethers.formatEther(shortfall)} + 10% buffer), have ${ethers.formatEther(balance)} AIUS`
+          `(shortfall: ${ethers.formatEther(shortfall)} + ${BLOCKCHAIN.STAKE_BUFFER_PERCENT}% buffer), have ${ethers.formatEther(balance)} AIUS`
         );
       }
 
@@ -244,6 +247,13 @@ export class BlockchainService implements IBlockchainService {
     }
   }
 
+  /**
+   * Submit a task to the Arbius network
+   * @param modelId - The model ID (bytes32 hash)
+   * @param input - JSON string of task input parameters
+   * @param fee - Additional fee to pay in addition to model fee (in wei)
+   * @returns The task ID of the submitted task
+   */
   async submitTask(modelId: string, input: string, fee: bigint): Promise<string> {
     const bytes = ethers.hexlify(ethers.toUtf8Bytes(input));
     const modelFee = (await this.arbius.models(modelId)).fee;
@@ -329,6 +339,11 @@ export class BlockchainService implements IBlockchainService {
     }
   }
 
+  /**
+   * Submit a solution for a task (includes commitment signaling)
+   * @param taskid - The task ID to submit solution for
+   * @param cid - The IPFS CID of the solution (hex format with 0x prefix)
+   */
   async submitSolution(taskid: string, cid: string): Promise<void> {
     const commitment = generateCommitment(this.wallet.address, taskid, cid);
     log.debug(`Generated commitment: ${commitment}`);
@@ -341,7 +356,7 @@ export class BlockchainService implements IBlockchainService {
     }
 
     // Sleep to avoid nonce issues
-    await new Promise(r => setTimeout(r, 1000));
+    await new Promise(r => setTimeout(r, BLOCKCHAIN.COMMITMENT_DELAY_MS));
 
     // Submit solution
     try {
@@ -388,7 +403,7 @@ export class BlockchainService implements IBlockchainService {
     try {
       const filter = this.arbius.filters.TaskSubmitted(taskid);
       const currentBlock = await this.provider.getBlockNumber();
-      const fromBlock = Math.max(0, currentBlock - 10000);
+      const fromBlock = Math.max(0, currentBlock - BLOCKCHAIN.BLOCK_LOOKBACK);
 
       log.debug(`Searching for taskid ${taskid} from block ${fromBlock} to ${currentBlock}`);
 
@@ -434,10 +449,18 @@ export class BlockchainService implements IBlockchainService {
     }
   }
 
+  /**
+   * Get the Arbius contract instance for direct interaction
+   * @returns The ethers Contract instance for Arbius
+   */
   getArbiusContract(): Contract {
     return this.arbius;
   }
 
+  /**
+   * Get the ethers provider instance
+   * @returns The FallbackProvider used for blockchain queries
+   */
   getProvider(): ethers.FallbackProvider {
     return this.provider;
   }
